refactor: MyAccoutBox - MyLikesBox 리팩토링 및 로딩 상태 추가
